backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Pydantic for data validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Google Gemini AI
google-generativeai==0.3.2

# Document processing
python-docx==1.1.0
PyPDF2==3.0.1
weasyprint==60.2

# Utilities
python-dateutil==2.8.2
uuid6==2024.1.12

# CORS
starlette==0.35.1
